LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting-api/reporting-api-works-on-frame-ancestors.https.sub-expected.txt



FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
FAIL No report to parent. assert_equals: CSP report sent, but not expecting one. expected 0 but got 1
